Fix npm install instructions
The instructions to `npm install grunt --save-dev` and so on for each dependency
are incorrect (even though it works).

That command is for installing a dependency (locally, inside
`./node_modules`) and (with the `--save-dev` flag) writing that dependency to the
`package.json` file.  So this can be used the first time you add a new dependency,
but isn't meant to be used again every time.

Because the dependencies are listed in `package.json` all you need to do to install
them is `npm install`.  You also never need admin rights to do this because it
installs them to `./node_modules` (npm will only install globally with the -g flag,
which is only meant for installing executables, not project dependencies). :)